cpp journey without testing

+++ cpp00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: ooulcaid <ooulcaid@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/22 22:11:56 by ooulcaid          #+#    #+#             */
-/*   Updated: 2024/06/23 13:18:43 by ooulcaid         ###   ########.fr       */
+/*   Updated: 2024/06/26 14:18:12 by ooulcaid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void    put_one(std::string arg)
 {
-    for (int i = 0; i < arg.length(); i++)
+    for (size_t i = 0; i < arg.length(); i++)
     {
         if (arg[i] >= 'a' && arg[i] <= 'z')
             arg[i] -= 32;
